Sync'd Tapestry/VCI data definitions, broke Tapestry
