+++ src/modules/tapestry.py
@@ -1,14 +1,12 @@
 from glob import glob
 import json
 import os
-from urllib.request import urlretrieve
-from urllib.parse import urlparse
 
 from docker.types.services import EndpointSpec, ServiceMode
 
 
 class Tapestry():
-	def __init__(self, docker, config):
+	def __init__(self, docker, config, data_path):
 		self.docker = docker
 		self.config = config
 
@@ -17,33 +15,21 @@ class Tapestry():
 		self.tapestry_path = self.config['tapestry']['directory']
 		self.app_path = os.path.join(self.tapestry_path, 'app')
 		self.config_path = os.path.join(self.tapestry_path, 'config')
-		self.data_path = None
+		self.data_path = data_path
 
-		self.get_data()
 		self.set_config()
-
-	def get_data(self):
-		source_path = self.config['data']['source']
-
-		# Download the dataset if necessary to the target location
-		if urlparse(source_path).scheme != '':
-			target_path = os.path.abspath(self.config['data']['target'])
-
-			os.makedirs(os.path.dirname(target_path), exist_ok=True)
-			urlretrieve(source_path, target_path)
-
-			self.data_path = target_path
-		else:
-			self.data_path = os.path.abspath(source_path)
 
 	def set_config(self):
 		config_file = glob(os.path.join(self.config_path, '*.json'))[0]
 		with open(config_file, 'r', encoding='utf8') as _file:
 			tapestry_config = json.load(_file)
 
+			data_file = os.listdir(self.data_path)[0]
+			filename = os.path.split(data_file)[1]
+
 			# Set dataset-specific Tapestry configuration
-			tapestry_config['filename'] = f'/data/{os.path.split(self.data_path)[1]}'
-			tapestry_config['dimensions'] = self.config['data']['dimensions']
+			tapestry_config['filename'] = filename
+			tapestry_config['dimensions'] = self.config['tapestry']['dimensions']
 			tapestry_config['cameraPosition'] = [
 				0,
 				0,
@@ -62,14 +48,15 @@ class Tapestry():
 			env=['APP_DIR=/app'],
 			mode=ServiceMode(
 				mode='replicated',
-				replicas=self.config['cluster'].get('replicas', 0)
+				replicas=self.config['cluster'].get('replicas', 1)
 			),
 			mounts=[
 				f'{self.app_path}:/app:ro',
 				f'{self.config_path}:/config:ro',
-				f'{os.path.dirname(self.data_path)}:/data:ro'
+				f'{self.data_path}:/data:ro'
 			],
-			name='tapestry'
+			name='tapestry',
+			networks=['substrate-tapestry-net']
 		)
 
 	def stop(self):
